absracting repo init and adding content
